lemonbar_wrapper: fix memory leak
If any realloc's fail, the previous blocks of memory will be left
allocated, and return_window_status will be set to NULL, and we'll lose
the address of the previous allocated blocks of memory since we set it
to NULL. Create a new pointer and set it to return_window_status before
reallocing, so we can check if realloc failed and then free the old
blocks of memory before erroring.

NOTE: this is most likely COMPLETELY unnecessary for this small, simple
program. However, it was a nice learning experience, so let's add it
anyways ;)

Signed-off-by: shagbag913 <sh4gbag913@gmail.com>